+++ rust/routee-compass-core/src/algorithm/search/search_algorithm.rs
@@ -3,8 +3,8 @@ use super::search_error::SearchError;
 use super::MinSearchTree;
 use crate::algorithm::search::search_algorithm_type::SearchAlgorithmType;
 use crate::model::frontier::frontier_model::FrontierModel;
-use crate::model::graph::graph_struct::Graph;
-use crate::model::graph::{edge_id::EdgeId, vertex_id::VertexId};
+use crate::model::road_network::graph::Graph;
+use crate::model::road_network::{edge_id::EdgeId, vertex_id::VertexId};
 use crate::model::termination::termination_model::TerminationModel;
 use crate::model::traversal::traversal_model::TraversalModel;
 use crate::util::read_only_lock::ExecutorReadOnlyLock;
