+++ rust/routee-compass-core/src/model/mod.rs
@@ -1,6 +1,6 @@
 pub mod cost;
 pub mod frontier;
-pub mod graph;
+pub mod road_network;
 pub mod property;
 pub mod termination;
 pub mod traversal;
