+++ rust/routee-compass-core/src/model/road_network/graph.rs
@@ -1,6 +1,6 @@
 use crate::algorithm::search::direction::Direction;
-use crate::model::graph::graph_error::GraphError;
-use crate::model::graph::{edge_id::EdgeId, vertex_id::VertexId};
+use crate::model::road_network::graph_error::GraphError;
+use crate::model::road_network::{edge_id::EdgeId, vertex_id::VertexId};
 use crate::model::property::edge::Edge;
 use crate::model::property::vertex::Vertex;
 use std::collections::HashMap;
