+++ rust/routee-compass-core/src/model/road_network/mod.rs
@@ -2,6 +2,6 @@ pub mod edge_id;
 pub mod edge_loader;
 pub mod graph_error;
 pub mod graph_loader;
-pub mod graph_struct;
+pub mod graph;
 pub mod vertex_id;
 pub mod vertex_loader;
