+++ rust/routee-compass-core/src/model/road_network/vertex_id.rs
