+++ rust/routee-compass-core/src/model/road_network/vertex_loader.rs
@@ -1,6 +1,6 @@
 use std::path::PathBuf;
 
-use crate::model::graph::graph_error::GraphError;
+use crate::model::road_network::graph_error::GraphError;
 use crate::model::property::vertex::Vertex;
 use crate::util::fs::read_utils;
 use kdam::{Bar, BarExt};
