+++ rust/routee-compass/src/app/compass/config/compass_configuration_error.rs
@@ -1,7 +1,7 @@
 use config::ConfigError;
 use routee_compass_core::{
     model::{
-        graph::graph_error::GraphError, traversal::traversal_model_error::TraversalModelError,
+        road_network::graph_error::GraphError, traversal::traversal_model_error::TraversalModelError,
     },
     util::conversion::conversion_error::ConversionError,
 };
